requirements.txt
anthropic
ashr-labs
PyMuPDF
openpyxl
python-dotenv
